Refactor exonum-java::node_builder -> exonum-java::node
